requirements.txt
redis>=4.5.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
duckdb>=0.9.0
neo4j>=5.8.0
playwright>=1.35.0
aiohttp>=3.8.4
langchain>=0.0.200
anthropic>=0.5.0
openai>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pytest>=7.3.1
pytest-asyncio>=0.21.0
mcp[cli]>=1.9.0